fix: 제안 견적 금액이 너무 크면 정체불명 에러로 실패하던 문제 수정
estimatedPrice가 Integer라 21억 넘는 값이 오면 JSON 파싱 단계에서 죽어
원인을 알 수 없는 실패로 보였다. Long으로 받아 파싱은 무사히 통과시키고
@Max(10억 미만)로 제대로 된 검증 메시지가 뜨게 했다. 프론트도 실제 실패
사유를 보여주도록 함께 수정.

+++ backend/post-service/src/main/java/com/team2/postservice/proposal/service/ProposalService.java
@@ -52,7 +52,7 @@ public class ProposalService {
         Proposal proposal = Proposal.builder()
                 .post(post)
                 .repairerEmail(repairerEmail)
-                .estimatedPrice(request.estimatedPrice())
+                .estimatedPrice(request.estimatedPrice().intValue())
                 .content(request.content())
                 .attachResume(Boolean.TRUE.equals(request.attachResume()))
                 .build();
